Task manager changes
Added edit task function, display task function. This allows the user to edit the task detail. display task just allows the user to see the details of the task they are editing. attempted to move the login functionality to a function but it defined to many global variables so I kept it in main.